\name{dcdflib-package}
\alias{dcdflib-package}
\alias{dcdflib}
\docType{package}
\title{
\packageTitle{dcdflib}
}
\description{
\packageDescription{dcdflib}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dcdflib}
\packageIndices{dcdflib}
%% ~~ An overview of how to use the package, including the most important ~~
%% ~~ functions ~~
}
\author{
\packageAuthor{dcdflib}

Maintainer: \packageMaintainer{dcdflib}
}
%% \references{
%% ~~ Literature or other references for background information ~~
%% }
\seealso{
  The package \code{\link[DPQ:DPQ-package]{DPQ}}, which is also about
  computations for DPQ -- Density, Probability, and Quantile -- functions.
}
%% \examples{
TODO
%% }
\keyword{ distribution }
\keyword{ package }
